Make msvc happy. Tests do not work for some reason though.

+++ benchmark/benchmark.cpp
@@ -11,14 +11,14 @@ constexpr auto SMALL_KNOWN_STRING = "Hello, World!";
 constexpr auto LARGE_KNOWN_STRING = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
 
 template <typename StringType>
-std::vector<StringType> generate_random_strings(size_t count, size_t min_length, size_t max_length, uint32_t seed)
+std::vector<StringType> generate_random_strings(size_t count, uint32_t min_length, uint32_t max_length, uint32_t seed)
 {
     std::vector<StringType> strings;
     strings.reserve(count);
     std::string alphabet = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%^&*()_+[]{}|;:,.<>?";
     std::mt19937 generator(seed);
     std::uniform_int_distribution<> length_distribution(min_length, max_length);
-    std::uniform_int_distribution<> char_distribution(0, alphabet.size() - 1);
+    std::uniform_int_distribution<> char_distribution(0, static_cast<int>(alphabet.size() - 1));
 
     constexpr float small_known_string_probability = 0.1f;  // 10% chance to generate the small known string
     constexpr float large_known_string_probability = 0.05f; // 5% chance to generate the large known string
@@ -52,9 +52,9 @@ std::vector<StringType> generate_random_strings(size_t count, size_t min_length,
 void StdStringCountExisting(benchmark::State &state)
 {
     size_t count = state.range(0);
-    size_t min_length = state.range(1);
-    size_t max_length = state.range(2);
-    uint32_t seed = state.range(3);
+    uint32_t min_length = static_cast<uint32_t>(state.range(1));
+    uint32_t max_length = static_cast<uint32_t>(state.range(2));
+    uint32_t seed = static_cast<uint32_t>(state.range(3));
     auto strings = generate_random_strings<std::string>(count, min_length, max_length, seed);
     for (auto _ : state)
     {
@@ -84,13 +84,13 @@ BENCHMARK(StdStringCountExisting)
 void GermanStringCountExisting(benchmark::State &state)
 {
     size_t count = state.range(0);
-    size_t min_length = state.range(1);
-    size_t max_length = state.range(2);
-    uint32_t seed = state.range(3);
+    uint32_t min_length = static_cast<uint32_t>(state.range(1));
+    uint32_t max_length = static_cast<uint32_t>(state.range(2));
+    uint32_t seed = static_cast<uint32_t>(state.range(3));
     auto strings = generate_random_strings<gs::german_string>(count, min_length, max_length, seed);
 
-    const gs::german_string small_known_string(SMALL_KNOWN_STRING, std::strlen(SMALL_KNOWN_STRING), gs::persistent_t{});
-    const gs::german_string large_known_string(LARGE_KNOWN_STRING, std::strlen(LARGE_KNOWN_STRING), gs::persistent_t{});
+    const gs::german_string small_known_string(SMALL_KNOWN_STRING, static_cast<uint32_t>(std::strlen(SMALL_KNOWN_STRING)), gs::persistent_t{});
+    const gs::german_string large_known_string(LARGE_KNOWN_STRING, static_cast<uint32_t>(std::strlen(LARGE_KNOWN_STRING)), gs::persistent_t{});
 
     for (auto _ : state)
     {
@@ -120,9 +120,9 @@ template <typename TString>
 void TemplateStringSort(benchmark::State &state)
 {
     size_t count = state.range(0);
-    size_t min_length = state.range(1);
-    size_t max_length = state.range(2);
-    uint32_t seed = state.range(3);
+    uint32_t min_length = static_cast<uint32_t>(state.range(1));
+    uint32_t max_length = static_cast<uint32_t>(state.range(2));
+    uint32_t seed = static_cast<uint32_t>(state.range(3));
     auto strings = generate_random_strings<TString>(count, min_length, max_length, seed);
 
     for (auto _ : state)
